Responsive design
I added some code to make sure all the elements are displaying correctly on all screen sizes
I had some issues with the cards on screens over 1200px but less than around 1600px. I decided to add a new breakpoint in order to make them appear correctly.
I also worked on the card with Emma's journey as I have removed the text on small screens and it did not show up on the big screen. After I added a new breakpoint I made some changes to its styling to make it appear with the text on screens above 1260px